Fix for user’s not having expected roles
The user has no roles object because it is the message
user instead of the user stored in the brain with
the roles associated with it.

To recreate the bug give your user admin role and
role a command role (ex. Foo). Limit a command to
foo.  At line 83 in your code:
``` robot.auth.hasRole(reqUser, opts.roles)```
was always evaluating to false because the user
Passed into the hasRole never had roles on it.

Hubot-Auth expects the user passed into that
function to have roles.

For reference please read code in hubot-auth at
Line 120
https://github.com/hubot-scripts/hubot-auth/blob/master/src/auth.coffee#L120